Use yaml_rust for BookOptions
`BookOptions` now internally uses the yaml_rust library. This should
not change much for an end-user perspective, but should make easy to
allow to configure a `Book` with a YAML block inside a markdown file.

Since the "old" Crowbooks's config format was similar, but somewhat
different, to markdown, this might cause some bugs,
though. Particularly, strings should now be escaped with "" in some
casess (e.g. if it contains special characters). On the other hand, it
*allows* to optionally espace a string with these quotes, which wasn't
possible until then and might be useful in some cases.